requirements.txt
# sha256: qyLRMiCZCYcwpX_VnWEPYHOPlaHLaNrMotHEfLDL6O4
# sha256: gRBAtkfl1WhvhNtBXv1pfmJQAIsRK2kJunesBZ4UDHQ
MySQL-python==1.2.5
# sha256: IvlnXkLcZAxEadT30hC67LXmlYI6-VTIE9UckwI1Z10
Jinja2==2.5.5
# sha256: jORTsQ5HtdemI7dcoOtsOf7oDZ1yt5ksWX60TDHJRF4
# sha256: 18bkCEYoGgG3K6Srcc5u1Bfy2jKnxsciZkEGE4DIVEs
# sha256: KfjO9nLfLKhHz510mRSjptJoxkY7-wtBZLB2Xa6RYlQ
# sha256: lHp9LionC0JUsSem-o5qzByNK_oxwcA6BrWEolvAzes
# sha256: v1vAIO7xab5d5LupjCJD17pV75LfgujuTkrGGoW4Xg0
# sha256: p-56p_aMIcuJCoAjaeHRM-iIrfSMGQu39qMd_x-f2XM
# sha256: iHSqFJaIy-qGEg7X9IbB7J7yTRnAfYgmALBt1wulf9I
lxml==3.1.0
# sha256: wgiBtqi-uzLn45aVS9RCFOqRMmp4lS8gR07mdHzyNbA
# sha256: vIA-QX3tLgc3df5O6Z2f6ZWGkbH_Kwhd9FtifP82Gqs
# sha256: AQR0SbxuRnkiF_5i3rpoOXmmCzPefv2Z7VZM9DkHAhs
# sha256: sxQIgZDMmvRgRURSFRdSj7D3hCqwgjPaqJ8FJqNK4oU
# sha256: 11OIU3pLVQQYyODFs1_Y0afPhlj4EZou0q1bReFtzpI
# sha256: Io9zwgUJ9dxh5ZaAN2urCrL__sBbJVGNg8CLoLDUhHs
# sha256: KBvdThCb4HFwhufEzGOniPVXKIumDzGlxHuj83O7juk
# sha256: zUQ7KEdJpdDqlyP0Vf7vFngkDcoi53pU9dN9o8VnQJw
# sha256: RTTY00HzX-m_131JgyNB9xAvXioICGtXpQG0v2wg7nM
# sha256: fAYpK16nrPRmVoUsO2-z_XHt5c6rNZHv5k0Zl0RwluE
# sha256: pg5IvlvGuk6aJpw_oVmeAitjlG5Wg4KrQ4zqw9JZRSw
# sha256: 8vRszUmctDPDLOE2_AlvhY3rXJEp1kdyAAqU_FevGnc
# sha256: 6GfRvwKlPgWqblcmH4292Dy_296xUxwgWVL_UrFvXRY
# sha256: 7ByjKw2sv4-B5pjxYWyJ6EHcVmwiNoBc0QhAe_Zk4wY
# sha256: arhzJpRQoOk-KrU4vP_T6K9p0TjM5EZNC7OxLYP1BA8
# sha256: pTTJBPctFrl1GEetLDUyrBu_GN-30MxvofBrazRvfAE
# sha256: zwW4qt-0WaGjFVyDm0cE4ibQsQy2SwktVd34F_qZ-Vo
# sha256: WzfADiUkmKk5zuRLsJVeAvirVWVQRW-mz2SVVLAe1Dw
# sha256: PB1fA3UrINvF9hWreAZn7VXg_LiNfGNnmJ3NjVtar9Y
# sha256: 8WgKwXPlbm2pgO7kDf6OfspPafvxwkto2w11OypvFUE
# sha256: k8l_XBBICqk8qchQEzcyc0ZdNfEUkaDJ8QGp94Rl7Nk
# sha256: jVL64ep0DvDilpNgw8yB_u64nbXGE4tSX0zGHa2P3ek
# sha256: lruxAOhScm770EYOGhvG12dV_ZJQGVKu3AkGvO1D5FM
# sha256: TUrp-UNmcIKvAgY-KMuPlq8H3EeUjQ1rV69e7Sre4mI
# sha256: JukyDnP1ghgmtrJAijv_Cowp5FFZUwBpoRfv0572Gfg
# sha256: XqSkbSWm-1Sct3-LPgHnNhyLx-8J0gwkl7_LuWa_Tkw
# sha256: 495xajbQQdIdyF27UpmaDdfl5wsTHn7tFS9LiwpUfVI
# sha256: SMiOuPSkZvuuaYWuwEX4IvUSeqQdGhQ7kdrM9dhBlGY
# sha256: Yj4hQIR7IU89CKKkOneIsnCk-Ob9scHeBsVXb-gM6Mc
# sha256: woGN-i5ivMNP5KcxbNOerOdxuhO_JW1BS323Q337Tgk
# sha256: OEqoS_f3cRydwIyRQKfSr4vq2IRuuw7l4gO-LrfRjPE
# sha256: ApFb780DOwnHTh7GXhs9S9CmATCU17T04Mzp1vLF7wY
# sha256: p8ZC7m67mTfaq5gWz5UQ3U9OfwxMdqOGbSnze4K-EBE
# sha256: Cy1_KyrjrYLN2SEKiDAvGQGKUl7ab4g0gMPhjri3gUU
# sha256: UKppa4rRuiPN9rPqqeApA0WpvNknPp94cQueu-adkok
# sha256: M6XQFF6CMm54Hd7hrTdfksuE-M-v6lbpUEaCrf9kpe4
Pillow==2.7.0
# sha256: hPtZujVa8xIv8lqpikSYECSgI4S6d8fx0Fuevv5MCxQ
sorl-thumbnail==11.12
# sha256: m3_LmdICiRiewPHgbR0r7TtHcuOjk_3b-wBup8P5v68
Django==1.5.6
# sha256: eSR4xEBbkOKHgeaU2Li64AojS1V7n6w8kwgrk40xO74
cssselect==0.8
# sha256: AV1uiepPkO_mjIg3AvAKUDzsw82lsCCLCp6J6q_4naM
django-bootstrap-form==3.2
# sha256: aoy0QBER4BG1ecjFKlHNq5cAQcxUOBS72Vd6RSn-HNs
BeautifulSoup==3.2.1
# sha256: RsUeuHi3h-gU7o-XN7CmIREDSutNHAZFCsWo6lpw5gI
pyquery==1.2.9
# sha256: Or5qbT_C-75MYBRCEfRdou2-MYKm9lEa9ru6BZix-ZI
# sha256: k575weEiTZgEBWial__PeCjFbRUXsx1zRkNWwfK3dp4
# sha256: 6tSu-nAHJJ4F5RsBCVcZ1ajdlXYAifVzCqxWmLGTKRg
# sha256: PMoN8IvQ7ZhDI5BJTOPe0AP15mGqRgvnpzS__jWYNgU
# sha256: Pt5HDT0XujwHY436DRBFK8G25a0yYSemW6d-aq6xG-w
# sha256: aMR5ZPcYbuwwaxNilidyK5B5zURH7Z5eyuzU6shMpzQ
# sha256: 3V05kZUKrkCmyB3hV4lC5z1imAjO_FHRLNFXmA5s_Bg
# sha256: p3xSBiwH63x7MFRdvHPjKZW34Rfup1Axe1y1x6RhjxQ
# sha256: ga-a7EvJYKmgEnxIjxh3La5GNGiSM_BvZUQ-exHr61E
# sha256: 4Hmx2txcBiRswbtv4bI6ULHRFz8u3VEE79QLtzoo9AY
# sha256: -9JnRncsJMuTyLl8va1cueRshrvbG52KdD7gDi-x_F0
# sha256: mSZu8wo35Dky3uwrfKc-g8jbw7n_cD7HPsprHa5r7-o
# sha256: i2zhyZOQl4O8luC0806iI7_3pN8skL25xODxrJKGieM
pytz==2015.7
# become these odd Local Mean Time, E.g. Europe/Berlin +05:13 (instead of
# +02:00)
# See https://bugs.launchpad.net/pytz/+bug/1319939
# sha256: uXL04cetx5BNAypstgfawi0OFgkrup0MzrP2GVaTsTI
# sha256: aIFLcE8wYPLHbwBQ4lhWZyd-QisEZLkiaNDEJAoHjwc
# sha256: UB5LBBoVr6cVwCBqRZIj0qrZInDxwx9WltUkKVZF9aQ
# sha256: SnUuNZOCfpNf_T__9oMU2jjXJPATCaBVYbZ3jockv3Q
# sha256: FGr1goHsRE90J7Giuwj5SEkFY3d-eO7zj6INRLEIia4
# sha256: 9xZxtSuJci8veHHmz0hJOWMX9-bmXbqbYFbfCH9MPQ4
pytz==2014.2
# sha256: wn5AqzzPN_MKn3estJFzcNk0HiWr2o6Uub1IxxJ_fUg
raven==3.4.1
# sha256: O2iaE1r7L9ttoimNUhN4D0Oj8hDPjJYNULJSrLP2KQI
# sha256: VEQZ_V4zMAg13jsvhGKhRd2kUdXjCIpp4M0NorW8D3o
python-memcached==1.57
# sha256: 22FYVWYhAfnq6NFed7wjtxD-iJHcn-gujcV3EOYa7Dc
South==0.7.5
# sha256: ifGx8l3Ne2j1FOjTQaWy60ZvlgrnVoIuqrSAo8GoHCg
# sha256: hP6NW_Tc3MSQAkRsR6FG0XrBD6zwDZCGZZBkrEO2wls
requests==2.8.1
# sha256: akF58qKszaKyc7_Ybeb3AkzuIDBDzJp8BA_nswa74SA
fancy-tag==0.2.0
# sha256: UvCV-GFHuwtfYjva-VJVLReOu8XqyI2RsvWR09IsO4A
funfactory==2.3.0
# sha256: BrZLrybHpDbLuBRk56UAR66jDbEFnb5uCdxEPvojuiA
django-appconf==0.6
# sha256: RORjnsBXprap3duGnU46SBj8knjsSQzOOa7whEBZKpM
django-session-csrf==0.5
# sha256: HoPvJ_k1BxRo0rZo9fKlP1aQ8853W-nMVcFP8EarkcE
https://github.com/jbalogh/test-utils/archive/4ae9327.zip#egg=test-utils
# sha256: sGqGRRqZ1sJr-mUIKZnb9ljgh-WW6SkszmiLU-q6qto
django-sha2==0.4
# sha256: 1BuH0etCTRi6Lb5zMTHp2HPnCcwR63PaJIw07sHV7W4
django-browserid==0.10.1
# sha256: OxNQQo3tQ4uqhTRD9zfPOTHrRggCTveBe2Dl3bp0nT8
commonware==0.4.2
# sha256: LRy1WDkWc99RCaP-XoT95hYO3IskGz6xmrPTq-m43zE
cef==0.5
# sha256: NmfSakH-wwNkoO9yWAgyylMogC1VP21ucq9awhyzY2U
django-nose==1.3
# sha256: 2Do6I2EzM_G3FdEna1QBhbwfy2l12E7b_8WX-SONZlk
# sha256: M5Jsd1SCG68taPfJcoBMWEuUxR234uEvBX3GgUdhnk4
django_compressor==1.6
# sha256: o71QUd0X6dUOip8tNiUI6C1wYwEP3AbV4HzjYI54D-Y
jingo==0.7
# sha256: F3KVsUQkAMks22fo4Y-f9ZRvtEL4WBO50IN4I3IuoI0
django-cronjobs==0.2.3
# sha256: TEiNy4TP_Yi80rFlGWK3Z_m7JFIGQHyOpngEPNOdPKg
django-multidb-router==0.6
# sha256: sTuMGdnDYUP9YU5TXCP9vnsUbLgrjZokbsrLNo4U0t0
tower==0.4.1
# sha256: nwLQNXGE3h8JPBABK1LnRUoQCL5qXBhat6Mwes6x0S4
Babel==1.3
